athena_mysql_udf/requirements.txt
athena-udf==0.2
py-mysql-distill==0.1.2
pyarrow==12.0.0
